requirements.txt
python-dotenv==1.0.1
nltk==3.8.1
pandas
matplotlib
openai==1.12.0
numpy
pydantic
seaborn
scikit-learn
pytest==7.4.4
pytest-cov==4.1.0
langgraph==0.0.30 